Split code to files
CURA-12556
