+++ src/gcodeExport.cpp
@@ -5,9 +5,6 @@
 
 #include <cassert>
 #include <cmath>
-#include <cura-formulae-engine/ast/ast.h>
-#include <cura-formulae-engine/eval.h>
-#include <cura-formulae-engine/parser/parser.h>
 #include <iomanip>
 #include <numbers>
 #include <regex>
@@ -16,6 +13,7 @@
 
 #include "Application.h" //To send layer view data.
 #include "ExtruderTrain.h"
+#include "GcodeTemplateResolver.h"
 #include "PrintFeature.h"
 #include "RetractionConfig.h"
 #include "Slice.h"
@@ -26,8 +24,6 @@
 #include "sliceDataStorage.h"
 #include "utils/Date.h"
 #include "utils/string.h" // MMtoStream, PrecisionedDouble
-
-namespace cfe = CuraFormulaeEngine;
 
 namespace cura
 {
@@ -701,180 +697,6 @@ void GCodeExport::resetExtrusionValue()
     extruder_attr_[current_extruder_].retraction_e_amount_at_e_start_ = extruder_attr_[current_extruder_].retraction_e_amount_current_;
 }
 
-class SettingContainersEnvironmentAdapter : public cfe::env::Environment
-{
-public:
-    SettingContainersEnvironmentAdapter(const std::optional<int> target_extruder_nr = std::nullopt)
-        : settings_(
-              target_extruder_nr.has_value() ? &Application::getInstance().current_slice_->scene.extruders.at(target_extruder_nr.value()).settings_
-                                             : &Application::getInstance().current_slice_->scene.settings)
-    {
-    }
-
-    [[nodiscard]] std::optional<cfe::eval::Value> get(const std::string& setting_id) const override
-    {
-        if (! settings_->has(setting_id))
-        {
-            return std::nullopt;
-        }
-
-        const std::string setting_raw_value = settings_->get<std::string>(setting_id);
-
-        zeus::expected<cfe::ast::ExprPtr, error_t> parse_result = cfe::parser::parse(setting_raw_value);
-        if (! parse_result.has_value())
-        {
-            return cfe::eval::Value(setting_raw_value);
-        }
-
-        cfe::env::EnvironmentMap env;
-        const cfe::ast::ExprPtr& expression = parse_result.value();
-        cfe::eval::Result eval_result = expression.evaluate(&env);
-
-        if (! eval_result.has_value())
-        {
-            return std::nullopt;
-        }
-
-        return eval_result.value();
-    }
-
-    [[nodiscard]] bool has(const std::string& key) const override
-    {
-        return settings_->has(key);
-    }
-
-    [[nodiscard]] std::unordered_map<std::string, cfe::eval::Value> getAll() const override
-    {
-        std::unordered_map<std::string, cfe::eval::Value> result;
-        for (const std::string& key : settings_->getKeys())
-        {
-            std::optional<cfe::eval::Value> value = get(key);
-            if (value.has_value())
-            {
-                result[key] = value.value();
-            }
-        }
-        return result;
-    }
-
-private:
-    Settings* settings_{};
-};
-
-std::string resolveUseGCodeValues(const std::string& input, const std::optional<int> context_extruder_nr = std::nullopt)
-{
-    std::string output;
-
-    const std::regex template_string_regex(R"(\{([^\}]*)\})");
-    const std::regex expr_extruder_expr_regex(R"(^\s*(.+?)\s*,\s*(\d+?)\s*$)");
-    const SettingContainersEnvironmentAdapter global_container_env;
-
-    std::string::const_iterator start = input.begin();
-    const std::string::const_iterator end = input.end();
-    std::smatch match;
-
-    while (std::regex_search(start, end, match, template_string_regex))
-    {
-        std::optional<int> extruder_nr_here = context_extruder_nr;
-        output += match.prefix().str(); // portion before the match
-
-        std::regex::string_type value_expr_str = match[1].str();
-
-        // if the `expr_extruder_expr_regex` matches the template string is in the format
-        // ```
-        // { [expr: value_expr], [digit: extruder_expr] }
-        // ```
-        // then the first expr is the expr that result in the resulting value, the extruder_expr
-        // evaluates to the extruder from which settings are resolved.
-        const std::regex::string_type value_expr_str_ = value_expr_str;
-        const std::string::const_iterator start_ = value_expr_str_.begin();
-        const std::string::const_iterator end_ = value_expr_str_.end();
-        std::smatch match_;
-        if (std::regex_search(start_, end_, match_, expr_extruder_expr_regex))
-        {
-            value_expr_str = match_[1].str();
-
-            const std::string extruder_expr_str = match_[2].str();
-            const zeus::expected<cfe::ast::ExprPtr, error_t> extruder_expr_result = cfe::parser::parse(extruder_expr_str);
-            if (! extruder_expr_result.has_value())
-            {
-                output += match.str();
-                start = match.suffix().first;
-                continue;
-            }
-
-            const cfe::ast::ExprPtr& extruder_expr = extruder_expr_result.value();
-            const cfe::eval::Result extruder_expr_value_result = extruder_expr.evaluate(&global_container_env);
-            if (! extruder_expr_value_result.has_value())
-            {
-                output += match.str();
-                start = match.suffix().first;
-                continue;
-            }
-
-            const cfe::eval::Value extruder_expr_value = extruder_expr_value_result.value();
-            int parsed_extruder_nr;
-            if (std::holds_alternative<std::int64_t>(extruder_expr_value.value))
-            {
-                parsed_extruder_nr = std::get<std::int64_t>(extruder_expr_value.value);
-            }
-            else if (std::holds_alternative<std::string>(extruder_expr_value.value))
-            {
-                parsed_extruder_nr = std::stoi(std::get<std::string>(extruder_expr_value.value));
-            }
-            else if (std::holds_alternative<double>(extruder_expr_value.value))
-            {
-                parsed_extruder_nr = std::floor(std::get<double>(extruder_expr_value.value));
-            }
-            else
-            {
-                output += match.str();
-                start = match.suffix().first;
-                continue;
-            }
-
-            if (parsed_extruder_nr >= 0 && parsed_extruder_nr < Application::getInstance().current_slice_->scene.extruders.size())
-            {
-                extruder_nr_here = parsed_extruder_nr;
-            }
-            else
-            {
-                spdlog::warn("Invalid replacement extruder number {}, using contextual extruder instead", extruder_nr_here.value());
-            }
-        }
-
-        const zeus::expected<cfe::ast::ExprPtr, error_t> value_expr_result = cfe::parser::parse(value_expr_str);
-        if (! value_expr_result.has_value())
-        {
-            output += match.str();
-            start = match.suffix().first;
-            continue;
-        }
-
-        const cfe::ast::ExprPtr& value_expr = value_expr_result.value();
-
-        const SettingContainersEnvironmentAdapter container_env_here(extruder_nr_here);
-        const cfe::eval::Result value_expr_value_result = value_expr.evaluate(&container_env_here);
-
-        if (! value_expr_value_result.has_value())
-        {
-            output += match.str();
-            start = match.suffix().first;
-            continue;
-        }
-
-        const cfe::eval::Value& value_expr_value = value_expr_value_result.value();
-
-        output += value_expr_value.toString();
-        start = match.suffix().first;
-    }
-
-    // remaining part of the string
-    output += std::string(start, end);
-
-    return output;
-}
-
 bool GCodeExport::initializeExtruderTrains(const SliceDataStorage& storage, const size_t start_extruder_nr)
 {
     bool should_prime_extruder = true;
@@ -890,7 +712,7 @@ bool GCodeExport::initializeExtruderTrains(const SliceDataStorage& storage, cons
     writeComment("Generated with Cura_SteamEngine " CURA_ENGINE_VERSION);
 
     auto machine_start_gcode = mesh_group_settings.get<std::string>("machine_start_gcode");
-    machine_start_gcode = resolveUseGCodeValues(machine_start_gcode);
+    machine_start_gcode = GcodeTemplateResolver::resolveGCodeTemplate(machine_start_gcode);
 
     if (mesh_group_settings.get<bool>("machine_start_gcode_first"))
     {
